pyproject.toml
[project]
name = "cis-app"
version = "0.1.0"
description = "A lightweight RAG LLM chat UI leveraging AWS S3 Vectors and OpenAI API"
authors = []
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "streamlit>=1.25.0",
  "numpy>=1.24.4",
  "openai>=1.82.1",
  "boto3>=1.39.17",
  "transformers>=4.52.4",
  "sentence-transformers>=4.1.0",
  "faiss-cpu>=1.10.0",
  "html2text",
  "humanfriendly",
]
